refactor: Change the mock repository to a public class
Reuse the class to the unit test Getting Balance
